move object from Services to Objects folder
